Revert: drop the two zero-impact commits
Per a measurement pass, two commits had zero (or negative) effect on the hello-world prod bundle:

- a0b1f09d7f (Move Meta mixin methods to standalone fns): bundle went from 134.17 KB → 134.19 KB (+0.02 KB). Mixin.create chain was already being tree-shaken in prod regardless.
- 75761b841e (Decouple VM debug symbols/names from opcodes.ts): bundle held at 134.12 KB. LOCAL_DEBUG=false in dist/prod (and dist/dev) was already constant-folding the debug branches out, and vite was already tree-shaking the unused @glimmer/debug imports out of the smoke-test bundle.

Both refactors were architecturally cleaner but pure no-ops at the bundle measurement that motivated this PR.
